Refactor OptionsSelector into its own component.
