+++ src/screens/internal/CompareSnapshots/CompareSnapshots.style.js
@@ -5,10 +5,3 @@ export const Wrapper = styled.div`
   padding: 1rem 2rem;
   min-height: calc(100vh - 64px);
 `;
-
-export const ModelSelectorContainer = styled.div`
-  background-color: white;
-  border: 1px solid black;
-  border-radius: 3px;
-  padding: 1rem;
-`;
